Move GossipState out of metrics
