+++ src/control/actor/gossip/topic/join.go
@@ -7,12 +7,13 @@ import (
 	"github.com/protolambda/rumor/control/actor/base"
 	"github.com/protolambda/rumor/metrics"
 	"github.com/protolambda/rumor/p2p/gossip"
+	"github.com/protolambda/rumor/control/actor/gossipimport"
 )
 
 type TopicJoinCmd struct {
 	*base.Base
 	PeerStore *metrics.PeerStore
-	GossipState   *metrics.GossipState
+	GossipState   *gossipimport.GossipState
 	Eth2TopicName string `ask:"--eth-topic" help:"The name of the eth2 topics"`
 	ForkDigest    string `ask:"--fork-version" help:"The fork digest value of the network we want to join to (Default Mainnet)"`
 	Encoding      string `ask:"--encoding" help:"Encoding that is getting used"`
